Create 344. Reverse String.js
Here the main logic is in the swapping part, initially it may seems that s[i]=s[j] and vice versa, but the catch is here we will lost the primary value, so need to use a temp variable for swapping. Another catch is the loop, if we run it length times then it will reverse and then again reverse, as a result we will get a output same as original one. So, need to run the loop half of the length of the array