Create 121. Best Time to Buy and Sell Stock.js
Here first catch that minBuy should not initialize with 0, with should initialize with 1st element or something very large. Then just need to compare and update values. Another catch that, it might seems brute force is needed but the catch is to make profit they need to buy at lowest price, so optimization works...